# src/main/java/org/groupx/emergencyresponse/dto/CreateEmergencyReportDTO.java
package org.groupx.emergencyresponse.dto;

public class CreateEmergencyReportDTO {
    public String type;
    public String description;
    public Double locationLat;
    public Double locationLng;
    public String reporterId;
}