# src/main/java/org/groupx/emergencyresponse/dto/EmergencyResponseDTO.java
package org.groupx.emergencyresponse.dto;

public class EmergencyResponseDTO {
    public Long reportId;
    public String responderName;
    public String responderRole;
    public int etaMinutes;
}